Implement a Google API wrapper client
Add client
