+++ src/OpenAIDemo.Web/Startup/ServiceExtensions.cs
@@ -136,6 +136,7 @@ public static class ServiceExtensions
     {
         services.AddTransient<LoggingHttpMessageHandler>();
         services.AddTransient<GeocodeClient>();
+        services.AddTransient<GoogleApiClient>();
         services.AddRefitClient<IOpenAIClient>(
             IOpenAIClient.BaseAddress,
             new JsonSerializerOptions
